+++ stream-video-android-compose/src/main/kotlin/io/getstream/video/android/compose/utils/ImageUtils.kt
@@ -28,7 +28,7 @@ import androidx.compose.ui.platform.LocalContext
 import coil.compose.AsyncImagePainter
 import coil.compose.rememberAsyncImagePainter
 import coil.request.ImageRequest
-import io.getstream.video.android.compose.R
+import io.getstream.video.android.ui.common.R
 import io.getstream.video.android.compose.imageloading.LocalStreamImageLoader
 import kotlin.math.abs
