+++ stream-video-android-ui-common/src/main/res/values/colors.xml
