--- stream-video-android-ui-xml/src/main/res/values/colors.xml
@@ -1,69 +1,0 @@
-<?xml version="1.0" encoding="utf-8"?>
-<!--
-    Copyright (c) 2014-2022 Stream.io Inc. All rights reserved.
-
-    Licensed under the Stream License;
-    you may not use this file except in compliance with the License.
-    You may obtain a copy of the License at
-
-      https://github.com/GetStream/stream-chat-android/blob/main/LICENSE
-
-    Unless required by applicable law or agreed to in writing, software
-    distributed under the License is distributed on an "AS IS" BASIS,
-    WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
-    See the License for the specific language governing permissions and
-    limitations under the License.
--->
-<resources>
-    <!-- Text -->
-    <color name="stream_xml_text_color_primary">@color/stream_xml_black</color>
-    <color name="stream_xml_text_color_secondary">@color/stream_xml_grey</color>
-    <color name="stream_xml_text_color_hint">@color/stream_xml_grey</color>
-
-    <!-- Accents -->
-    <color name="stream_xml_accent_blue">#005FFF</color>
-    <color name="stream_xml_accent_red">#FF3742</color>
-    <color name="stream_xml_accent_green">#20E070</color>
-
-    <!-- Greys -->
-    <color name="stream_xml_black">#000000</color>
-    <color name="stream_xml_black_50">#80000000</color>
-    <color name="stream_xml_grey">#7A7A7A</color>
-    <color name="stream_xml_grey_gainsboro">#DBDBDB</color>
-    <color name="stream_xml_grey_whisper">#ECEBEB</color>
-    <color name="stream_xml_white_smoke">#F2F2F2</color>
-    <color name="stream_xml_white_snow">#FCFCFC</color>
-    <color name="stream_xml_white">#FFFFFF</color>
-    <color name="stream_xml_blue_alice">#E9F2FF</color>
-
-    <!-- Specials -->
-    <color name="stream_xml_overlay">#33000000</color>
-    <color name="stream_xml_overlay_dark">#99000000</color>
-    <color name="stream_xml_background_gradient_start">#F7F7F7</color>
-    <color name="stream_xml_background_gradient_end">#FCFCFC</color>
-    <color name="stream_xml_message_list_image_attachment_background">#E9EAED</color>
-    <color name="stream_xml_message_list_video_attachment_background">#E9EAED</color>
-
-    <!-- Other -->
-    <color name="stream_xml_icon_button_background">#FFFFFF</color>
-    <color name="stream_xml_button_background_color">#005AFF</color>
-    <color name="stream_xml_button_text_color">#FFFFFF</color>
-    <color name="stream_xml_border">#E8E8E8</color>
-    <color name="stream_xml_highlight">#FBF4DD</color>
-
-    <!-- Avatar gradient colors -->
-    <color name="stream_xml_avatar_gradient_orange">#FF8A65</color>
-    <color name="stream_xml_avatar_gradient_green">#81C784</color>
-    <color name="stream_xml_avatar_gradient_blue">#4FC3F7</color>
-    <color name="stream_xml_avatar_gradient_brown">#A1887F</color>
-    <color name="stream_xml_avatar_gradient_yellow">#FFF176</color>
-    <color name="stream_xml_avatar_gradient_red">#E57373</color>
-    <array name="stream_xml_avatar_gradient_colors">
-        <item>@color/stream_xml_avatar_gradient_orange</item>
-        <item>@color/stream_xml_avatar_gradient_green</item>
-        <item>@color/stream_xml_avatar_gradient_blue</item>
-        <item>@color/stream_xml_avatar_gradient_brown</item>
-        <item>@color/stream_xml_avatar_gradient_yellow</item>
-        <item>@color/stream_xml_avatar_gradient_red</item>
-    </array>
-</resources>
